+++ src/main/java/org/archive/extract/WATExtractorOutput.java
@@ -2,11 +2,13 @@ package org.archive.extract;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
+import java.io.File;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.nio.charset.Charset;
 import java.text.ParseException;
+import java.net.UnknownHostException;
 import java.util.Date;
 
 import org.archive.format.gzip.GZIPMemberWriter;
@@ -22,6 +24,12 @@ import org.archive.util.StreamCopy;
 import org.archive.util.io.CommitedOutputStream;
 import org.json.JSONException;
 
+import java.net.InetAddress;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+
+import java.util.logging.Logger;
+
 public class WATExtractorOutput implements ExtractorOutput {
 	WARCRecordWriter recW;
 	private boolean wroteFirst;
@@ -29,11 +37,15 @@ public class WATExtractorOutput implements ExtractorOutput {
 	private static int DEFAULT_BUFFER_RAM = 1024 * 1024;
 	private int bufferRAM = DEFAULT_BUFFER_RAM;
 	private final static Charset UTF8 = Charset.forName("UTF-8");
+	private String outputFile;
 	
-	public WATExtractorOutput(OutputStream out) {
+	private static final Logger LOG = Logger.getLogger(WATExtractorOutput.class.getName());
+	
+	public WATExtractorOutput(OutputStream out, String outputFile) {
 		gzW = new GZIPMemberWriter(out);
 		recW = new WARCRecordWriter();
 		wroteFirst = false;
+		this.outputFile = outputFile;
 	}
 
 	private CommitedOutputStream getOutput() {
@@ -56,9 +68,9 @@ public class WATExtractorOutput implements ExtractorOutput {
 			throw new IOException("Missing Envelope.Format");
 		}
 		cos = getOutput();
-		if(envelopeFormat.equals("ARC")) {
+		if(envelopeFormat.startsWith("ARC")) {
 			writeARC(cos,top);
-		} else if(envelopeFormat.equals("WARC")) {
+		} else if(envelopeFormat.startsWith("WARC")) {
 			writeWARC(cos,top);
 		} else {
 			// hrm...
@@ -68,13 +80,51 @@ public class WATExtractorOutput implements ExtractorOutput {
 	}
 
 	private void writeWARCInfo(OutputStream recOut, MetaData md) throws IOException {
-		String filename = JSONUtils.extractSingle(md, "Container.Filename");
-		if(filename == null) {
-			throw new IOException("No Container.Filename...");
+		// filename is given in the command line
+		String filename = outputFile;
+		if (filename == null || filename.length() == 0) {
+			// if no filename by command line, we construct a default filename base on container filename
+			filename = JSONUtils.extractSingle(md, "Container.Filename");
+			if (filename == null) {
+				throw new IOException("No Container.Filename...");
+			}
+			if (filename.endsWith(".warc") || filename.endsWith(".warc.gz")) {
+				filename = filename.replaceFirst("\\.warc$", ".warc.wat.gz");
+				filename = filename.replaceFirst("\\.warc\\.gz$", ".warc.wat.gz");
+			} else if (filename.endsWith(".arc") || filename.endsWith(".arc.gz")) {
+				filename = filename.replaceFirst("\\.arc$", ".arc.wat.gz");
+				filename = filename.replaceFirst("\\.arc\\.gz$", ".arc.wat.gz");
+			}
 		}
+		// removing path from filename
+		File tmpFile = new File(filename);
+		filename = tmpFile.getName();
 		HttpHeaders headers = new HttpHeaders();
-		headers.add("Software-Info", IAUtils.COMMONS_VERSION);
-		headers.addDateHeader("Extracted-Date", new Date());
+		headers.add("software", IAUtils.COMMONS_VERSION);
+		headers.addDateHeader("extractedDate", new Date());
+
+		// add ip, hostname
+		try {
+			InetAddress host = InetAddress.getLocalHost();
+			headers.add("ip", host.getHostAddress());
+			headers.add("hostname", host.getCanonicalHostName());
+		} catch (UnknownHostException e) {
+			LOG.warning("unable to obtain local crawl engine host :\n"+e.getMessage());
+        }
+
+		headers.add("format", IAUtils.WARC_FORMAT);
+		headers.add("conformsTo", IAUtils.WARC_FORMAT_CONFORMS_TO);
+		// optional arguments
+		if(IAUtils.OPERATOR != null && IAUtils.OPERATOR.length() > 0) {
+			headers.add("operator", IAUtils.OPERATOR);
+		}
+		if(IAUtils.PUBLISHER != null && IAUtils.PUBLISHER.length() > 0) {
+			headers.add("publisher", IAUtils.PUBLISHER);
+		}
+		if(IAUtils.WAT_WARCINFO_DESCRIPTION != null && IAUtils.WAT_WARCINFO_DESCRIPTION.length() > 0) {
+			headers.add("description", IAUtils.WAT_WARCINFO_DESCRIPTION);
+		}
+		
 		ByteArrayOutputStream baos = new ByteArrayOutputStream();
 		headers.write(baos);
                 recW.writeWARCInfoRecord(recOut,filename,baos.toByteArray());
@@ -105,8 +155,9 @@ public class WATExtractorOutput implements ExtractorOutput {
 		} else {
 			targetURI = extractOrIO(md, "Envelope.WARC-Header-Metadata.WARC-Target-URI");
 		}
-		String capDateString = extractOrIO(md, "Envelope.WARC-Header-Metadata.WARC-Date");
-		capDateString = transformWARCDate(capDateString);
+		// handle date of generation in WARC format
+		DateFormat dateFormat = new SimpleDateFormat("yyyyMMddHHmmss");
+		String capDateString = dateFormat.format(new Date());
 		String recId = extractOrIO(md, "Envelope.WARC-Header-Metadata.WARC-Record-ID");
 		writeWARCMDRecord(recOut,md,targetURI,capDateString,recId);
 	}