+++ CHANGES.md
@@ -1,3 +1,7 @@
+1.1.7
+-----
+* [Store origin-code of ARC file header](https://github.com/iipc/webarchive-commons/pull/52/)
+
 1.1.6
 -----
 * [Handle empty String argument in CharsetDetector.trimAttrValue](https://github.com/iipc/webarchive-commons/pull/49)